Platform Class (basically what a section had)
Every platform prefab has certain important information:

1. Anchors: the bounds of the platform used to generate connecting platforms.
2. Spawn Location: Right now there is one per platform - this will be used for the player spawn, A collectable, or an enemy.
3. Width: Used in special types of platforms (Moving, rotating)
