+++ Assets/Prefabs/BuildingBlocks/2x0.5.prefab
@@ -104,6 +104,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 5008762508460989841}
+  - component: {fileID: 2211210796338924526}
   m_Layer: 9
   m_Name: 2x0.5
   m_TagString: Untagged
@@ -126,3 +127,19 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &2211210796338924526
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5008762508460989840}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 640a62a263600734d9d816b8cd4148e2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  spawnPoint: {x: 1, y: 1, z: 0}
+  leftAnchor: {x: 0, y: 0, z: 0}
+  rightAnchor: {x: 2, y: 0, z: 0}
+  width: 2
