+++ Assets/Prefabs/BuildingBlocks/3x1.prefab
@@ -104,6 +104,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 6693775090792741149}
+  - component: {fileID: 2545679372913815924}
   m_Layer: 9
   m_Name: 3x1
   m_TagString: Untagged
@@ -126,3 +127,19 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &2545679372913815924
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5736057811185428383}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 640a62a263600734d9d816b8cd4148e2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  spawnPoint: {x: 1.5, y: 1, z: 0}
+  leftAnchor: {x: 0, y: 0, z: 0}
+  rightAnchor: {x: 3, y: 0, z: 0}
+  width: 3
