+++ Assets/Prefabs/BuildingBlocks/3x2.prefab
@@ -9,6 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 8747290735714752853}
+  - component: {fileID: 4917320732930615095}
   m_Layer: 9
   m_Name: 3x2
   m_TagString: Untagged
@@ -31,6 +32,22 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &4917320732930615095
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3086496886497976997}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 640a62a263600734d9d816b8cd4148e2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  spawnPoint: {x: 1.5, y: 1, z: 0}
+  leftAnchor: {x: 0, y: 0, z: 0}
+  rightAnchor: {x: 3, y: 0, z: 0}
+  width: 3
 --- !u!1 &4073364161203599392
 GameObject:
   m_ObjectHideFlags: 0
