+++ Assets/Scripts/Platforms.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8b16002b1d460bb44b25e3efbda43310
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
