+++ Assets/Scripts/Platforms/Platform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 640a62a263600734d9d816b8cd4148e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
